requirements.txt

viam-sdk==0.52.0

typing-extensions
